Rename api client proxy
