--- apps/public/src/apiClient.ts
@@ -1,6 +1,0 @@
-import * as apiClient from '@meldingen/api-client'
-
-// Set base url for the API client
-apiClient.OpenAPI.BASE = 'http://localhost:8000'
-
-export * from '@meldingen/api-client'
