+++ apps/public/src/apiClientProxy.ts
@@ -1,0 +1,6 @@
+import * as apiClientProxy from '@meldingen/api-client'
+
+// Set base url for the API client
+apiClientProxy.OpenAPI.BASE = 'http://localhost:8000'
+
+export * from '@meldingen/api-client'
