+++ apps/public/src/app/(general)/bijlage/_components/FileUpload.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import type { ChangeEvent, InputHTMLAttributes } from 'react'
 
-import type { AttachmentOutput } from 'apps/public/src/apiClient'
-import { postMeldingByMeldingIdAttachment } from 'apps/public/src/apiClient'
+import type { AttachmentOutput } from 'apps/public/src/apiClientProxy'
+import { postMeldingByMeldingIdAttachment } from 'apps/public/src/apiClientProxy'
 
 import { useMeldingContext } from '../../../../context/MeldingContextProvider'
